Add files via upload
The "Temperature Converter" is a graphical user interface (GUI) application designed to assist users in converting temperature values between Celsius, Fahrenheit, and Kelvin units. Developed using the tkinter library in Python, this application provides a straightforward and user-friendly interface for performing temperature conversions accurately and effortlessly.

Upon launching the application, a window titled "Temperature Converter" appears. The window dimensions are set to 500 pixels in width and 250 pixels in height, providing an optimal display area for the user interface components.